volttron-ven/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
boto3>=1.28.0,<2.0.0

# Hardware interface dependencies
pyserial>=3.5
RPi.GPIO>=0.7.1
